improve error test cov and re-arrange for more straightfwd testing
