+++ Makefile
@@ -1,10 +1,6 @@
-build:
-	@cargo build
 
-test:
-	@cargo test --manifest-path serde-tests/Cargo.toml
-	@cargo test
-
-clean:
-	@cargo clean --manifest-path serde-tests/Cargo.toml
-	@cargo clean
+test-coverage:
+	@docker run -it --rm \
+		--security-opt seccomp=unconfined \
+		-v "$(PWD):/volume" \
+		xd009642/tarpaulin
