--- serde-tests/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "serde-tests"
-version = "0.1.0"
-authors = ["Doug Tangren <d.tangren@gmail.com>"]
-
-[dependencies]
-serde = "1.0"
-serde_derive = "1.0"
-envy = { path = ".." }
-
-[lib]
-name = "serde_tests"
-path = "lib.rs"
-
-[[test]]
-name = "serde"
-path = "test.rs"
