--- serde-tests/lib.rs
@@ -1,1 +1,0 @@
-// blank
